LPS-73734 Move upgrade schema to a minor version
