+++ portal-workflow-kaleo-service/bnd.bnd
@@ -3,6 +3,6 @@ Bundle-SymbolicName: com.liferay.portal.workflow.kaleo.service
 Bundle-Version: 2.0.30
 Liferay-Releng-Module-Group-Description:
 Liferay-Releng-Module-Group-Title: Kaleo Workflow Engine
-Liferay-Require-SchemaVersion: 1.3.4
+Liferay-Require-SchemaVersion: 1.4.1
 Liferay-Service: true
 -dsannotations-options: inherit
