+++ portal-workflow-kaleo-service/src/main/java/com/liferay/portal/workflow/kaleo/internal/upgrade/KaleoServiceUpgrade.java
@@ -24,7 +24,7 @@ import com.liferay.portal.workflow.kaleo.internal.upgrade.v1_3_0.UpgradeKaleoAct
 import com.liferay.portal.workflow.kaleo.internal.upgrade.v1_3_0.UpgradeKaleoDefinition;
 import com.liferay.portal.workflow.kaleo.internal.upgrade.v1_3_2.UpgradeKaleoClassNameAndKaleoClassPK;
 import com.liferay.portal.workflow.kaleo.internal.upgrade.v1_3_3.UpgradeBlogsEntryClassName;
-import com.liferay.portal.workflow.kaleo.internal.upgrade.v1_3_4.UpgradeKaleoDefinitionVersion;
+import com.liferay.portal.workflow.kaleo.internal.upgrade.v1_4_1.UpgradeKaleoDefinitionVersion;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -74,12 +74,15 @@ public class KaleoServiceUpgrade implements UpgradeStepRegistrator {
 
 		registry.register(
 			"com.liferay.portal.workflow.kaleo.service", "1.3.2", "1.3.3",
-			new UpgradeBlogsEntryClassName(),
-			new com.liferay.portal.workflow.kaleo.internal.upgrade.v1_3_3.
+			new UpgradeBlogsEntryClassName());
+
+		registry.register(
+			"com.liferay.portal.workflow.kaleo.service", "1.3.3", "1.4.0",
+			new com.liferay.portal.workflow.kaleo.internal.upgrade.v1_4_0.
 				UpgradeSchema());
 
 		registry.register(
-			"com.liferay.portal.workflow.kaleo.service", "1.3.3", "1.3.4",
+			"com.liferay.portal.workflow.kaleo.service", "1.4.0", "1.4.1",
 			new UpgradeKaleoDefinitionVersion());
 	}
